BUG: Fix DataParallel error
When running the network in DataParallel, the attention blocks are not replicated correctly, causing the forward pass to fail.
Fix this by not assigning the forward function dynamically in `operation_function()`.

Moreover, when the net is a DataParallel object, the `apply_argmax_softmax` function is not accessible from `self.net`, but from `self.net.module`.