ENH: Add reloading of (partially) finished hyperspace
If the hyperspace file already exists, load it and compare it to the generated hyperspace by matching on the parameter columns.
Then store the results of matching entries in self.results and rename folders to match new index (generated from the configuration file hyperspace).
Indices that are not matched get a '_' suffix.

Finally, only train and test the hyperspaces for which there was no result (yet).